requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.8.24
click==6.7
decorator==4.3.0
Django==2.1
Flask==1.0.2
future==0.16.0
gcloud==0.17.0
geocoder==1.38.1
googleapis-common-protos==1.5.3
httplib2==0.11.3
image==1.5.24
ipython==7.0.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jws==0.1.3
line-profiler==2.1.2
MarkupSafe==1.0
oauth2client==3.0.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==2.0.5
protobuf==3.6.0
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycryptodome==3.4.3
Pygments==2.2.0
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2018.5
ratelim==0.1.6
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
selenium==3.12.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
